+++ portal/home/models_ui_text.py
@@ -36,13 +36,15 @@ class _BlockNames:
     rich_text = _('RichText')
     title = _('Title')
     line = _('Line')
+    chart = _('Chart')
 
 
 # Streamfield blocks' `help_text`s.
 class _StreamfieldHelpTexts:
     title_block = _('''
-    A title that displays full-width. The WebPage's body should start with this
-    title. The "Small Title" shows above the "Big Title".
+    A title that displays full-width. The WebPage's body should start with
+    a title block. The title shows on a green background, or on a provided
+    background image. The "Small Title" shows above the "Big Title".
     ''')
     rich_text = _('''
     Text that allows rich-formatted content: headers; bold and italic text;
@@ -64,8 +66,12 @@ class _StreamfieldHelpTexts:
     Many CardLinks can display in a responsive grid.
     ''')
     line_block = _('''
-    A full-width line that separates blocks of content and that shows as a
-    green line in the web page.
+    A full-width line that separates blocks of content and shows as a green
+    line on the web page.
+    ''')
+    chart_block = _('''
+    A Superset chart from MyEQIP's superset instance. The drop-down menu
+    lists available charts. The selected chart displays in the web page.
     ''')